refactor(deprize): split pure quote math for unit testing; guard router readiness
- Move betSlice/betBudget/buildAmounts/searchMaxQtyWithinCost into a
  dependency-free quote-math module so the mocha/ts-node unit runner can load
  them without the app's ESM deps; quote.ts and constants.ts re-export them
- Make the binary search truly maximal at the exact-budget boundary while still
  never overspending
- Show a loading state until router.isReady so the detail page doesn't flash an
  invalid-id notice on first client render
